data-loader/requirements.txt
trino==0.327.0
pandas==2.2.0
pyarrow==15.0.0
minio==7.2.3
python-dotenv==1.0.1
requests==2.31.0
sodapy==2.2.0
tabulate==0.9.0
matplotlib==3.8.2
seaborn==0.13.0
duckdb==0.9.2

# New Flask API dependencies
Flask>=2.2.2
flask-restx>=1.0.3
Werkzeug==2.2.2
flasgger>=0.9.5
gunicorn>=20.1.0